Implemented PausableThreadPoolExecutor and CommandFileWatcher
Changed the Manager to use a new PausableThreadPoolExecutor that
provides new methods pause() and resume().

Implemented a CommandFileWatcher that looks for the presence of a file,
path configured with the COMMAND_FILE option. If the file exists and
has been modified, it parses it as a properties file, looking for a
property with either of the commands [PAUSE,RESUME,STOP] and
THREAD-COUNT=x. The corresponding methods are invoked to pause(),
resume(), or stop() and to adjust the THREAD-COUNT as long as it is a
positive integer value.

This resolves #14 and resolves #15.