resolves issue #31
Added an implementation of usage() to AbstractManager that leverages a
Usage annotation on each Option to dynamically generate a Usage message
that describes each of the CoRB2 Options currently available. As new
options are added, the Usage message will be updated. Added a unit test
to ensure that all of the Options use the Usage annotation.