initial checkin of Java client driven jobs
Initial work to be able to use the Java client for corb jobs for issue
#61
- Attempted to keep DMSDK functionality in separate DataMovement*
classes
- The majority of changes to existing classes are to make it easier to
override or swap implementations (i.e. use Iterator instead of
ResultSequence)
- The only instance of Java client bleeding into existing corb classes
(I think) is in the AbstractTask.getValueAsBytes() methods. We may need
to use a custom Iterator and/or item that could be customized to avoid
this, but needed it so that existing tasks, such as ExportToFileTask
and ExportBatchToFileTask would work…