getAbsoluteFile before attempting to getParentFile to avoid potential NPE
When a File object is instantiated with a `null` parent parameter, getParentFile() will return `null`. We want to obtain the parent directory of the file object in order to create any necessary directories with `mkdir()`. We can avoid NPE by resolving the absolute file.

Moving this logic into the `getExportFile()` method, in order to remove duplication and ensure that it is consistently applied.

Resolves issue #85